Plan 9 support
Since Plan 9 does not have symlinks, these problems do not occur.
Therefore, SecureJoinVFS and SecureJoin can map to filepath.Join,
along with the test for rootpath containing ..

Split tests so some can run on Plan 9

Move common variables and functions into common.go
