+++ join_test.go
@@ -13,7 +13,6 @@ import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 // TODO: These tests won't work on plan9 because it doesn't have symlinks, and
@@ -24,64 +23,9 @@ type input struct {
 	expected     string
 }
 
-func expandedTempDir(t *testing.T) string {
-	dir := t.TempDir()
-	dir, err := filepath.EvalSymlinks(dir)
-	require.NoError(t, err)
-	return dir
-}
-
-// Test basic handling of symlink expansion.
-func TestSymlink(t *testing.T) {
-	dir := expandedTempDir(t)
-
-	symlink(t, "somepath", filepath.Join(dir, "etc"))
-	symlink(t, "../../../../../../../../../../../../../etc", filepath.Join(dir, "etclink"))
-	symlink(t, "/../../../../../../../../../../../../../etc/passwd", filepath.Join(dir, "passwd"))
-
-	rootOrVol := string(filepath.Separator)
-	if vol := filepath.VolumeName(dir); vol != "" {
-		rootOrVol = vol + rootOrVol
-	}
-
-	tc := []input{
-		// Make sure that expansion with a root of '/' proceeds in the expected fashion.
-		{rootOrVol, filepath.Join(dir, "passwd"), filepath.Join(rootOrVol, "etc", "passwd")},
-		{rootOrVol, filepath.Join(dir, "etclink"), filepath.Join(rootOrVol, "etc")},
-
-		{rootOrVol, filepath.Join(dir, "etc"), filepath.Join(dir, "somepath")},
-		// Now test scoped expansion.
-		{dir, "passwd", filepath.Join(dir, "somepath", "passwd")},
-		{dir, "etclink", filepath.Join(dir, "somepath")},
-		{dir, "etc", filepath.Join(dir, "somepath")},
-		{dir, "etc/test", filepath.Join(dir, "somepath", "test")},
-		{dir, "etc/test/..", filepath.Join(dir, "somepath")},
-	}
-
-	for _, test := range tc {
-		got, err := SecureJoin(test.root, test.unsafe)
-		if err != nil {
-			t.Errorf("securejoin(%q, %q): unexpected error: %v", test.root, test.unsafe, err)
-			continue
-		}
-		// This is only for OS X, where /etc is a symlink to /private/etc. In
-		// principle, SecureJoin(/, pth) is the same as EvalSymlinks(pth) in
-		// the case where the path exists.
-		if test.root == "/" {
-			if expected, err := filepath.EvalSymlinks(test.expected); err == nil {
-				test.expected = expected
-			}
-		}
-		if got != test.expected {
-			t.Errorf("securejoin(%q, %q): expected %q, got %q", test.root, test.unsafe, test.expected, got)
-			continue
-		}
-	}
-}
-
 // In a path without symlinks, SecureJoin is equivalent to Clean+Join.
 func TestNoSymlink(t *testing.T) {
-	dir := expandedTempDir(t)
+	dir := t.TempDir()
 
 	tc := []input{
 		{dir, "somepath", filepath.Join(dir, "somepath")},
@@ -112,92 +56,6 @@ func TestNoSymlink(t *testing.T) {
 	}
 }
 
-// Make sure that .. is **not** expanded lexically.
-func TestNonLexical(t *testing.T) {
-	dir := expandedTempDir(t)
-
-	mkdirAll(t, filepath.Join(dir, "subdir"), 0o755)
-	mkdirAll(t, filepath.Join(dir, "cousinparent", "cousin"), 0o755)
-	symlink(t, "../cousinparent/cousin", filepath.Join(dir, "subdir", "link"))
-	symlink(t, "/../cousinparent/cousin", filepath.Join(dir, "subdir", "link2"))
-	symlink(t, "/../../../../../../../../../../../../../../../../cousinparent/cousin", filepath.Join(dir, "subdir", "link3"))
-
-	for _, test := range []input{
-		{dir, "subdir", filepath.Join(dir, "subdir")},
-		{dir, "subdir/link/test", filepath.Join(dir, "cousinparent", "cousin", "test")},
-		{dir, "subdir/link2/test", filepath.Join(dir, "cousinparent", "cousin", "test")},
-		{dir, "subdir/link3/test", filepath.Join(dir, "cousinparent", "cousin", "test")},
-		{dir, "subdir/../test", filepath.Join(dir, "test")},
-		// This is the divergence from a simple filepath.Clean implementation.
-		{dir, "subdir/link/../test", filepath.Join(dir, "cousinparent", "test")},
-		{dir, "subdir/link2/../test", filepath.Join(dir, "cousinparent", "test")},
-		{dir, "subdir/link3/../test", filepath.Join(dir, "cousinparent", "test")},
-	} {
-		got, err := SecureJoin(test.root, test.unsafe)
-		if err != nil {
-			t.Errorf("securejoin(%q, %q): unexpected error: %v", test.root, test.unsafe, err)
-			continue
-		}
-		if got != test.expected {
-			t.Errorf("securejoin(%q, %q): expected %q, got %q", test.root, test.unsafe, test.expected, got)
-			continue
-		}
-	}
-}
-
-// Make sure that symlink loops result in errors.
-func TestSymlinkLoop(t *testing.T) {
-	dir := expandedTempDir(t)
-
-	mkdirAll(t, filepath.Join(dir, "subdir"), 0o755)
-	symlink(t, "../../../../../../../../../../../../../../../../path", filepath.Join(dir, "subdir", "link"))
-	symlink(t, "/subdir/link", filepath.Join(dir, "path"))
-	symlink(t, "/../../../../../../../../../../../../../../../../self", filepath.Join(dir, "self"))
-
-	for _, test := range []struct {
-		root, unsafe string
-	}{
-		{dir, "subdir/link"},
-		{dir, "path"},
-		{dir, "../../path"},
-		{dir, "subdir/link/../.."},
-		{dir, "../../../../../../../../../../../../../../../../subdir/link/../../../../../../../../../../../../../../../.."},
-		{dir, "self"},
-		{dir, "self/.."},
-		{dir, "/../../../../../../../../../../../../../../../../self/.."},
-		{dir, "/self/././.."},
-	} {
-		got, err := SecureJoin(test.root, test.unsafe)
-		if !errors.Is(err, syscall.ELOOP) {
-			t.Errorf("securejoin(%q, %q): expected ELOOP, got %q & %v", test.root, test.unsafe, got, err)
-			continue
-		}
-	}
-}
-
-// Make sure that ENOTDIR is correctly handled.
-func TestEnotdir(t *testing.T) {
-	dir := expandedTempDir(t)
-
-	mkdirAll(t, filepath.Join(dir, "subdir"), 0o755)
-	writeFile(t, filepath.Join(dir, "notdir"), []byte("I am not a directory!"), 0o755)
-	symlink(t, "/../../../notdir/somechild", filepath.Join(dir, "subdir", "link"))
-
-	for _, test := range []struct {
-		root, unsafe string
-	}{
-		{dir, "subdir/link"},
-		{dir, "notdir"},
-		{dir, "notdir/child"},
-	} {
-		_, err := SecureJoin(test.root, test.unsafe)
-		if err != nil {
-			t.Errorf("securejoin(%q, %q): unexpected error: %v", test.root, test.unsafe, err)
-			continue
-		}
-	}
-}
-
 // Some silly tests to make sure that all error types are correctly handled.
 func TestIsNotExist(t *testing.T) {
 	for _, test := range []struct {
@@ -218,121 +76,6 @@ func TestIsNotExist(t *testing.T) {
 		got := IsNotExist(test.err)
 		if got != test.expected {
 			t.Errorf("IsNotExist(%#v): expected %v, got %v", test.err, test.expected, got)
-		}
-	}
-}
-
-type mockVFS struct {
-	lstat    func(path string) (os.FileInfo, error)
-	readlink func(path string) (string, error)
-}
-
-func (m mockVFS) Lstat(path string) (os.FileInfo, error) { return m.lstat(path) }
-func (m mockVFS) Readlink(path string) (string, error)   { return m.readlink(path) }
-
-// Make sure that SecureJoinVFS actually does use the given VFS interface.
-func TestSecureJoinVFS(t *testing.T) {
-	dir := expandedTempDir(t)
-
-	mkdirAll(t, filepath.Join(dir, "subdir"), 0o755)
-	mkdirAll(t, filepath.Join(dir, "cousinparent", "cousin"), 0o755)
-	symlink(t, "../cousinparent/cousin", filepath.Join(dir, "subdir", "link"))
-	symlink(t, "/../cousinparent/cousin", filepath.Join(dir, "subdir", "link2"))
-	symlink(t, "/../../../../../../../../../../../../../../../../cousinparent/cousin", filepath.Join(dir, "subdir", "link3"))
-
-	for _, test := range []input{
-		{dir, "subdir", filepath.Join(dir, "subdir")},
-		{dir, "subdir/link/test", filepath.Join(dir, "cousinparent", "cousin", "test")},
-		{dir, "subdir/link2/test", filepath.Join(dir, "cousinparent", "cousin", "test")},
-		{dir, "subdir/link3/test", filepath.Join(dir, "cousinparent", "cousin", "test")},
-		{dir, "subdir/../test", filepath.Join(dir, "test")},
-		// This is the divergence from a simple filepath.Clean implementation.
-		{dir, "subdir/link/../test", filepath.Join(dir, "cousinparent", "test")},
-		{dir, "subdir/link2/../test", filepath.Join(dir, "cousinparent", "test")},
-		{dir, "subdir/link3/../test", filepath.Join(dir, "cousinparent", "test")},
-	} {
-		var nLstat, nReadlink int
-		mock := mockVFS{
-			lstat:    func(path string) (os.FileInfo, error) { nLstat++; return os.Lstat(path) },
-			readlink: func(path string) (string, error) { nReadlink++; return os.Readlink(path) },
-		}
-
-		got, err := SecureJoinVFS(test.root, test.unsafe, mock)
-		if err != nil {
-			t.Errorf("securejoin(%q, %q): unexpected error: %v", test.root, test.unsafe, err)
-			continue
-		}
-		if got != test.expected {
-			t.Errorf("securejoin(%q, %q): expected %q, got %q", test.root, test.unsafe, test.expected, got)
-			continue
-		}
-		if nLstat == 0 && nReadlink == 0 {
-			t.Errorf("securejoin(%q, %q): expected to use either lstat or readlink, neither were used", test.root, test.unsafe)
-		}
-	}
-}
-
-// Make sure that SecureJoinVFS actually does use the given VFS interface, and
-// that errors are correctly propagated.
-func TestSecureJoinVFSErrors(t *testing.T) {
-	var (
-		lstatErr    = errors.New("lstat error")
-		readlinkErr = errors.New("readlink err")
-	)
-
-	dir := expandedTempDir(t)
-
-	// Make a link.
-	symlink(t, "../../../../../../../../../../../../../../../../path", filepath.Join(dir, "link"))
-
-	// Define some fake mock functions.
-	lstatFailFn := func(string) (os.FileInfo, error) { return nil, lstatErr }
-	readlinkFailFn := func(string) (string, error) { return "", readlinkErr }
-
-	// Make sure that the set of {lstat, readlink} failures do propagate.
-	for idx, test := range []struct {
-		vfs      VFS
-		expected []error
-	}{
-		{
-			expected: []error{nil},
-			vfs: mockVFS{
-				lstat:    os.Lstat,
-				readlink: os.Readlink,
-			},
-		},
-		{
-			expected: []error{lstatErr},
-			vfs: mockVFS{
-				lstat:    lstatFailFn,
-				readlink: os.Readlink,
-			},
-		},
-		{
-			expected: []error{readlinkErr},
-			vfs: mockVFS{
-				lstat:    os.Lstat,
-				readlink: readlinkFailFn,
-			},
-		},
-		{
-			expected: []error{lstatErr, readlinkErr},
-			vfs: mockVFS{
-				lstat:    lstatFailFn,
-				readlink: readlinkFailFn,
-			},
-		},
-	} {
-		_, err := SecureJoinVFS(dir, "link", test.vfs)
-
-		success := false
-		for _, exp := range test.expected {
-			if errors.Is(err, exp) {
-				success = true
-			}
-		}
-		if !success {
-			t.Errorf("SecureJoinVFS.mock%d: expected to get lstatError, got %v", idx, err)
 		}
 	}
 }
